Debug flag needs to be moved out to a separate package
If we don't do that any code that depends on it depends on root
package and it become imposible to evacuate such code to a separate
package.
